# Conflicts:
#	public/js/acheminements.js
#	public/js/script-wiilog.js
#	public/js/status.js
#	public/js/utilisateur.js
#	src/Controller/AcheminementsController.php
#	src/Controller/MouvementStockController.php
#	templates/acheminements/modalEditContentAcheminements.html.twig
#	templates/acheminements/modalNewContentAcheminements.html.twig
#	templates/mouvement_traca/newMassMvtTraca.html.twig
#	templates/status/modalEditStatusContent.html.twig
#	templates/status/modalNewStatus.html.twig
#	templates/utilisateur/modalEditUserContent.html.twig
#	templates/utilisateur/modalNewUser.html.twig

+++ src/Controller/StatusController.php
@@ -5,9 +5,12 @@ namespace App\Controller;
 
 use App\Entity\Action;
 use App\Entity\CategorieStatut;
+use App\Entity\CategoryType;
 use App\Entity\Menu;
 use App\Entity\Statut;
 
+use App\Entity\Type;
+use App\Service\StatusService;
 use App\Service\UserService;
 
 use Doctrine\ORM\EntityManagerInterface;
@@ -21,6 +24,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @Route("/statuts")
@@ -31,9 +35,15 @@ class StatusController extends AbstractController
      * @var UserService
      */
     private $userService;
+    private $statusService;
+    private $translator;
 
-    public function __construct(UserService $userService) {
+    public function __construct(UserService $userService,
+                                TranslatorInterface $translator,
+                                StatusService $statusService) {
         $this->userService = $userService;
+        $this->translator = $translator;
+        $this->statusService = $statusService;
     }
 
     /**
@@ -48,52 +58,40 @@ class StatusController extends AbstractController
         }
 
         $categoryStatusRepository = $entityManager->getRepository(CategorieStatut::class);
-		$categoriesStatusLitigeArr = $categoryStatusRepository->findByLabelLike('litige');
+        $typeRepository = $entityManager->getRepository(Type::class);
+        $categories = $categoryStatusRepository->findByLabelLike(
+            [
+                CategorieStatut::DISPATCH,
+                CategorieStatut::LITIGE_ARR,
+                CategorieStatut::LITIGE_RECEPT
+            ]
+        );
+		$types = $typeRepository->findByCategoryLabel(CategoryType::DEMANDE_DISPATCH);
+
+        $categoryStatusDispatchIds = array_filter($categories, function ($category) {
+            return $category['nom'] === CategorieStatut::DISPATCH;
+        });
 
         return $this->render('status/index.html.twig', [
-            'categories' => $categoriesStatusLitigeArr,
+            'categories' => $categories,
+            'categoryStatusDispatchId' => array_values($categoryStatusDispatchIds)[0]['id'] ?? 0,
+            'types' => $types
         ]);
     }
 
     /**
      * @Route("/api", name="status_param_api", options={"expose"=true}, methods="GET|POST")
      * @param Request $request
-     * @param EntityManagerInterface $entityManager
      * @return Response
      */
-    public function api(Request $request,
-                        EntityManagerInterface $entityManager): Response
+    public function api(Request $request): Response
     {
         if ($request->isXmlHttpRequest()) {
             if (!$this->userService->hasRightFunction(Menu::PARAM, Action::DISPLAY_STATU_LITI)) {
                 return $this->redirectToRoute('access_denied');
             }
 
-            $statusRepository = $entityManager->getRepository(Statut::class);
-
-            $listStatusLitigeArr = $statusRepository->findByCategorieName(CategorieStatut::LITIGE_ARR);
-            $listStatusLitigeRecep = $statusRepository->findByCategorieName(CategorieStatut::LITIGE_RECEPT);
-
-            $rows = [];
-            foreach (array_merge($listStatusLitigeArr, $listStatusLitigeRecep) as $status) {
-                $url['edit'] = $this->generateUrl('status_api_edit', ['id' => $status->getId()]);
-
-                $rows[] =
-                    [
-                        'Label' => $status->getNom(),
-                        'Categorie' => $status->getCategorie() ? $status->getCategorie()->getNom() : '',
-                        'Comment' => $status->getComment(),
-                        'Treated' => $status->isTreated() ? 'oui' : 'non',
-                        'NotifToBuyer' => $status->getSendNotifToBuyer() ? 'oui' : 'non',
-                        'NotifToDeclarant' => $status->getSendNotifToDeclarant() ? 'oui' : 'non',
-                        'Order' => $status->getDisplayOrder() ?? '',
-                        'Actions' => $this->renderView('status/datatableStatusRow.html.twig', [
-                            'url' => $url,
-                            'statusId' => $status->getId(),
-                        ]),
-                    ];
-            }
-            $data['data'] = $rows;
+            $data = $this->statusService->getDataForDatatable($request->request);
             return new JsonResponse($data);
         }
         throw new NotFoundHttpException("404");
@@ -102,12 +100,14 @@ class StatusController extends AbstractController
     /**
      * @Route("/creer", name="status_new", options={"expose"=true}, methods="GET|POST")
      * @param Request $request
+     * @param StatusService $statusService
      * @param EntityManagerInterface $entityManager
      * @return Response
      * @throws NoResultException
      * @throws NonUniqueResultException
      */
     public function new(Request $request,
+                        StatusService $statusService,
                         EntityManagerInterface $entityManager): Response
     {
         if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
@@ -117,41 +117,62 @@ class StatusController extends AbstractController
 
             $statutRepository = $entityManager->getRepository(Statut::class);
             $categoryStatusRepository = $entityManager->getRepository(CategorieStatut::class);
+            $typeRepository = $entityManager->getRepository(Type::class);
 
             // on vérifie que le label n'est pas déjà utilisé
             $labelExist = $statutRepository->countByLabelAndCategory($data['label'], $data['category']);
 
             if (!$labelExist) {
                 $category = $categoryStatusRepository->find($data['category']);
-                $status = new Statut();
-				$status
-                    ->setNom($data['label'])
-                    ->setComment($data['description'])
-					->setTreated($data['treated'])
-                    ->setSendNotifToBuyer($data['sendMails'])
-                    ->setSendNotifToDeclarant($data['sendMailsDeclarant'])
-					->setDisplayOrder((int)$data['displayOrder'])
-                    ->setCategorie($category);
+                $type = $typeRepository->find($data['type']);
+                $defaultForCategory = $data['defaultForCategory'] ?? false;
+                $statusCanBeCreated = (
+                    !$defaultForCategory
+                    || $statusService->canStatusBeDefault($entityManager, $category->getNom(), $type)
+                );
+                if (!$statusCanBeCreated) {
+                    $success = false;
+                    $message = 'Vous ne pouvez pas créer un statut par défaut pour cette entité et ce type, il en existe déjà un.';
+                }
+                else {
+                    $type = $typeRepository->find($data['type']);
+                    $status = new Statut();
+                    $status
+                        ->setNom($data['label'])
+                        ->setComment($data['description'])
+                        ->setTreated($data['treated'] ?? false)
+                        ->setDefaultForCategory($defaultForCategory)
+                        ->setSendNotifToBuyer((bool)$data['sendMails'])
+                        ->setSendNotifToDeclarant((bool)$data['sendMailsDeclarant'])
+                        ->setSendNotifToRecipient((bool)$data['sendMailsRecipient'])
+                        ->setNeedsMobileSync((bool)$data['needsMobileSync'])
+                        ->setDisplayOrder((int)$data['displayOrder'])
+                        ->setCategorie($category)
+                        ->setType($type ?? null);
 
-                $entityManager->persist($status);
-                $entityManager->flush();
+                    $entityManager->persist($status);
+                    $entityManager->flush();
 
-				return new JsonResponse([
-					'success' => true,
-					'msg' => 'Le statut "' . $data['label'] . '" a bien été créé.'
-				]);
+                    $success = true;
+                    $message = 'Le statut "' . $data['label'] . '" a bien été créé.';
+                }
             } else {
-				return new JsonResponse([
-					'success' => false,
-					'msg' => 'Le statut "' . $data['label'] . '" existe déjà pour cette catégorie. Veuillez en choisir un autre.'
-				]);
+                $success = false;
+                $message = 'Le statut "' . $data['label'] . '" existe déjà pour cette catégorie. Veuillez en choisir un autre.';
             }
+            return new JsonResponse([
+                'success' => $success,
+                'msg' => $message
+            ]);
         }
         throw new NotFoundHttpException("404");
     }
 
     /**
      * @Route("/api-modifier", name="status_api_edit", options={"expose"=true}, methods="GET|POST")
+     * @param Request $request
+     * @param EntityManagerInterface $entityManager
+     * @return Response
      */
     public function apiEdit(Request $request,
                             EntityManagerInterface $entityManager): Response
@@ -165,11 +186,30 @@ class StatusController extends AbstractController
             $categoryStatusRepository = $entityManager->getRepository(CategorieStatut::class);
 
             $status = $statutRepository->find($data['id']);
-            $categories = $categoryStatusRepository->findByLabelLike('litige');
+            $typeRepository = $entityManager->getRepository(Type::class);
+            $categories = $categoryStatusRepository->findByLabelLike([
+                CategorieStatut::DISPATCH,
+                CategorieStatut::LITIGE_ARR,
+                CategorieStatut::LITIGE_RECEPT
+            ]);
+            $types = $typeRepository->findByCategoryLabel(CategoryType::DEMANDE_DISPATCH);
+
+            $transCategories = array_map(
+                function (array $category) {
+                    return [
+                        'id' => $category['id'],
+                        'nom' => $category['nom'] === 'acheminement'
+                            ? $this->translator->trans('acheminement.acheminements')
+                            : $category['nom']
+                    ];
+                },
+                $categories
+            );
 
             $json = $this->renderView('status/modalEditStatusContent.html.twig', [
                 'status' => $status,
-                'categories' => $categories,
+                'categories' => $transCategories,
+                'types' => $types
             ]);
 
             return new JsonResponse($json);
@@ -180,10 +220,12 @@ class StatusController extends AbstractController
     /**
      * @Route("/modifier", name="status_edit",  options={"expose"=true}, methods="GET|POST")
      * @param EntityManagerInterface $entityManager
+     * @param StatusService $statusService
      * @param Request $request
      * @return Response
      */
     public function edit(EntityManagerInterface $entityManager,
+                         StatusService $statusService,
                          Request $request): Response
     {
         if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
@@ -192,38 +234,53 @@ class StatusController extends AbstractController
             }
 
             $statutRepository = $entityManager->getRepository(Statut::class);
-            $categoryStatusRepository = $entityManager->getRepository(CategorieStatut::class);
+            $typeRepository = $entityManager->getRepository(Type::class);
 
 			$status = $statutRepository->find($data['status']);
             $statusLabel = $status->getNom();
 
             // on vérifie que le label n'est pas déjà utilisé
-            $labelExist = $statutRepository->countByLabelDiff($data['label'], $statusLabel, $data['category']);
+            $categoryName = $status->getCategorie() ? $status->getCategorie()->getNom() : '';
+            $labelExist = $statutRepository->countByLabelDiff($data['label'], $statusLabel, $categoryName);
 
             if (!$labelExist) {
-                $category = $categoryStatusRepository->find($data['category']);
-                $status
-                    ->setNom($data['label'])
-                    ->setCategorie($category)
-					->setTreated($data['treated'])
-                    ->setSendNotifToBuyer($data['sendMails'])
-                    ->setSendNotifToDeclarant($data['sendMailsDeclarant'])
-					->setDisplayOrder((int)$data['displayOrder'])
-                    ->setComment($data['comment']);
+                $defaultForCategory = $data['defaultForCategory'] ?? false;
+                $statusCanBeCreated = (
+                    !$defaultForCategory
+                    || $statusService->canStatusBeDefault($entityManager, $categoryName, $status->getType(), $status)
+                );
+                if (!$statusCanBeCreated) {
+                    $success = false;
+                    $message = 'Vous ne pouvez pas ajouter de statut par défaut pour cette entité et ce type, il en existe déjà un.';
+                }
+                else {
+                    $type = $typeRepository->find($data['type']);
+                    $status
+                        ->setNom($data['label'])
+                        ->setTreated($data['treated'] ?? false)
+                        ->setDefaultForCategory($defaultForCategory)
+                        ->setSendNotifToBuyer((bool)$data['sendMails'])
+                        ->setSendNotifToDeclarant((bool)$data['sendMailsDeclarant'])
+                        ->setSendNotifToRecipient((bool)$data['sendMailsRecipient'])
+                        ->setNeedsMobileSync((bool)$data['needsMobileSync'])
+                        ->setDisplayOrder((int)$data['displayOrder'])
+                        ->setComment($data['comment'])
+                        ->setType($type ?? null);
 
-                $entityManager->persist($status);
-                $entityManager->flush();
+                    $entityManager->persist($status);
+                    $entityManager->flush();
 
-                return new JsonResponse([
-                	'success' => true,
-					'msg' => 'Le statut "' . $statusLabel . '" a bien été modifié.'
-				]);
+                    $success = true;
+                    $message = 'Le statut "' . $statusLabel . '" a bien été modifié.';
+                }
             } else {
-                return new JsonResponse([
-                	'success' => false,
-					'msg' => 'Le statut "' . $data['label'] . '" existe déjà pour cette catégorie. Veuillez en choisir un autre.'
-				]);
+                $success = false;
+                $message = 'Le statut "' . $data['label'] . '" existe déjà pour cette catégorie. Veuillez en choisir un autre.';
             }
+            return new JsonResponse([
+                'success' => $success,
+                'msg' => $message
+            ]);
         }
         throw new NotFoundHttpException('404');
     }
@@ -239,14 +296,14 @@ class StatusController extends AbstractController
     public function checkStatusCanBeDeleted(Request $request,
                                             EntityManagerInterface $entityManager): Response
     {
-        if ($request->isXmlHttpRequest() && $typeId = json_decode($request->getContent(), true)) {
+        if ($request->isXmlHttpRequest() && $statusId = json_decode($request->getContent(), true)) {
             if (!$this->userService->hasRightFunction(Menu::PARAM, Action::DELETE)) {
                 return $this->redirectToRoute('access_denied');
             }
 
             $statutRepository = $entityManager->getRepository(Statut::class);
 
-            $statusIsUsed = $statutRepository->countUsedById($typeId);
+            $statusIsUsed = $statutRepository->countUsedById($statusId);
 
             if (!$statusIsUsed) {
                 $delete = true;
@@ -266,6 +323,8 @@ class StatusController extends AbstractController
      * @param EntityManagerInterface $entityManager
      * @param Request $request
      * @return Response
+     * @throws NoResultException
+     * @throws NonUniqueResultException
      */
     public function delete(EntityManagerInterface $entityManager,
                            Request $request): Response
@@ -278,6 +337,13 @@ class StatusController extends AbstractController
             $statutRepository = $entityManager->getRepository(Statut::class);
 
             $status = $statutRepository->find($data['status']);
+            $statusIsUsed = $statutRepository->countUsedById($status->getId());
+            if ($statusIsUsed) {
+                return new JsonResponse([
+                    'success' => false,
+                    'msg' => 'Ce statut est utilisé, vous ne pouvez pas le supprimer.'
+                ]);
+            }
 
             $entityManager->remove($status);
             $entityManager->flush();
